requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
motor==2.5.0
pydantic==1.8.2
python-decouple==3.5